backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=0.19.1
openai>=0.27.0
langchain-openai>=0.0.267
langchain-pinecone>=0.0.267
langchain>=0.0.267
langchain-community>=0.0.267
bs4>=0.0.1
requests>=2.27.1
alembic>=1.7.4
psycopg2-binary>=2.9.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
sqlalchemy-utils>=0.38.2
aiofiles>=0.7.0
tenacity>=8.0.1
numpy>=1.21.2
tiktoken>=0.3.0
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.2
